+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [?] - 2025-07-??
 
+### Changed
+
+- **Breaking:** updated Bag-validation result format to match profile-validation 
+
 ### Added
 
 - added BagIt profile- and Bag-validation functionality